requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
docker>=6.0.0
pyyaml>=6.0
click>=8.0.0
requests>=2.31.0
GitPython>=3.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
